Created a Form component and added all the form code to it from my Contact page in order to make the app more modular and reusable.
